fix 커뮤니티 수정

+++ src/main/java/com/kosta/geekku/service/InteriorSeviceImpl.java
@@ -9,6 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -150,25 +151,25 @@ public class InteriorSeviceImpl implements InteriorService {
 		Interior interior = interiorRepository.findByCompany_companyId(companyId);
 		sample.setCompany(company);
 		sample.setInterior(interior);
-
-		if (coverImage != null && !coverImage.isEmpty()) {
-			// 파일 저장 경로 확인 및 디렉토리 생성
-			File uploadDir = new File(uploadPath);
-			if (!uploadDir.exists()) {
-				uploadDir.mkdirs(); // 디렉토리 생성
-			}
-
-			String fileName = coverImage.getOriginalFilename();
-			String filePath = uploadPath + "sampleImage/" + fileName;
-
-			File file = new File(filePath);
-			coverImage.transferTo(file);
-			sample.setCoverImage(file.getName());
-			interiorSampleRepository.save(sample);
-		}
-
+		
+        if (coverImage != null && !coverImage.isEmpty()) {
+        	// 파일 저장 경로 확인 및 디렉토리 생성
+            File uploadDir = new File(uploadPath);
+            if (!uploadDir.exists()) {
+                uploadDir.mkdirs();  // 디렉토리 생성
+            }
+            
+            String fileName = coverImage.getOriginalFilename();
+            String filePath = uploadPath + "sampleImage/" + fileName;
+                        
+            File file = new File(filePath);
+            coverImage.transferTo(file);
+            sample.setCoverImage(file.getName()); 
+            interiorSampleRepository.save(sample);
+        }
+		
 		interiorSampleRepository.save(sample);
-
+		
 		return sample.getSampleNum();
 	}
 
@@ -240,11 +241,12 @@ public class InteriorSeviceImpl implements InteriorService {
 
 	@Transactional
 	@Override
-	public Integer interiorRequest(String userId, InteriorRequestDto requestDto) throws Exception {
+	public Integer interiorRequest(String userId,InteriorRequestDto requestDto) throws Exception {
 		InteriorRequest request = requestDto.toEntity();
 		User user = User.builder().userId(UUID.fromString(userId)).build();
-		Interior interior = Interior.builder().interiorNum(1).build(); // test용 interiorNum 1 대입
-
+		Interior interior = Interior.builder().interiorNum(1).build();	//test용 interiorNum 1 대입
+		
+		
 		request.setUser(user);
 		request.setInterior(interior);
 		interiorRequestRepository.save(request);
@@ -308,8 +310,7 @@ public class InteriorSeviceImpl implements InteriorService {
 	}
 
 	@Override
-	public Integer updateReview(ReviewDto reviewDto, Integer num, List<Integer> delFileNum,
-			List<MultipartFile> fileList) throws Exception {
+	public Integer updateReview(ReviewDto reviewDto, Integer num, List<Integer> delFileNum, List<MultipartFile> fileList) throws Exception {
 		InteriorReview review = interiorReviewRepository.findById(num)
 				.orElseThrow(() -> new Exception("인테리어 후기 글번호 오류"));
 
@@ -318,22 +319,21 @@ public class InteriorSeviceImpl implements InteriorService {
 		review.setLocation(reviewDto.getLocation());
 		review.setStyle(reviewDto.getStyle());
 		review.setType(reviewDto.getType());
-
+		
 		interiorReviewRepository.save(review);
-
+		
 		// 기존 이미지파일 삭제하는 경우
-		if (delFileNum != null) {
-			for (Integer fn : delFileNum) {
+		if (delFileNum != null) {			
+			for (Integer fn: delFileNum) {				
 				File oldFile = new File(uploadPath, fn + "");
-				if (oldFile != null)
-					oldFile.delete();
+				if (oldFile != null) oldFile.delete();
 				interiorReviewImageRepository.deleteById(fn);
 			}
 		}
-
+				
 		// 이미지파일 추가
 		if (fileList != null && fileList.size() > 0) {
-			for (MultipartFile file : fileList) {
+			for (MultipartFile file: fileList) {
 				InteriorReviewImage bFile = new InteriorReviewImage();
 				bFile.setDierctory(uploadPath);
 				bFile.setName(file.getOriginalFilename());
@@ -341,12 +341,12 @@ public class InteriorSeviceImpl implements InteriorService {
 				bFile.setContentType(file.getContentType());
 				bFile.setInteriorReview(review);
 				interiorReviewImageRepository.save(bFile);
-
+						
 				File nFile = new File(uploadPath, bFile.getInteriorReviewImageNum() + "");
 				file.transferTo(nFile);
 			}
 		}
-
+				
 		return review.getReviewNum();
 	}
 
@@ -448,11 +448,9 @@ public class InteriorSeviceImpl implements InteriorService {
 		interiorRepository.save(interior);
 		return new HashMap<String, Object>(interiorDto.getInteriorNum());
 
-	}
 
 	public ReviewDto getReview(Integer reviewNum) throws Exception {
-		InteriorReview review = interiorReviewRepository.findById(reviewNum)
-				.orElseThrow(() -> new Exception("인테리어 후기 글번호 오류"));
+		InteriorReview review = interiorReviewRepository.findById(reviewNum).orElseThrow(() -> new Exception("인테리어 후기 글번호 오류"));
 		return review.toDto();
 
 	}
